Tie up loose ends with piecewise-linear arrival rate estimation.
Now supports fixing points on either the left or right side of the window of estimation, or both. This is useful as it is not always possible to estimate every segment due to low arrival counts. What we're able to do to counter this problem is to set (for example) the rate in that segment comstant, and continue estimating the next segment (if possible) with a piecewise linear function, where we fix one point.